Fight entropy a little harder
When the user undoes an action, Sublime will adjust the plugin-owned
regions to their previous values.  However, sometimes undoing results
in a change to the proven region.  In those cases, Sublime's helpful
region reversion is wrong.  To combat that case, this commit has the
plugin update the regions at every opportunity.