Make the display thread-safe and more efficient
This change improves the overall thread safety of the display classes
(see source code comment for rationale).  It also enables batching for
display updates, so frequent updates do not cause enormous strain on
the editor.  The CoqWorker threads tend to make many frequent display
updates, so this change has a big impact.