Listen to every change on every buffer
This fixes #9.

After consideration and reading some recommendations [1] on the
official ST issue tracker, I have decided this is the best approach.
It saves us from overwrought (and likely bug-ridden) schemes for
manually attaching/detaching TextChangeEventListeners at the right
times.

As the source code comment states, we can always switch to the more
complicated scheme if need arises.  For now, however, this approach
seems cheap enough to be practical.

[1]: https://github.com/sublimehq/sublime_text/issues/3536#issuecomment-670506712