Resolve settings using view-speific settings first
This fixes a number of weird behaviors and latent bugs surrounding use
of `Settings.set(...)`.  There doesn't seem to be a way to make such
an impure approach ironclad; it is better to adjust how settings are
read than to modify the in-memory settings object.

This commit fixes the following strange/buggy behaviors:

  1. When a project-specific setting is removed from the project
     definition, the plugin would not go back to using the global
     setting.
  2. Project-specific overrides for dynamic settings like "show_goals"
     would be ignored until restarting the Coq session.
  3. Project-specific overrides for dynamic settings like "show_goals"
     could be forgotten if the global settings were modified while
     Coq was running.

(There's a lesson in here somewhere about the benefits of using a more
complex read algorithm rather than trying to keep the system state as
one where the shortest possible read algorithm behaves correctly.
Modifying the in-memory settings to make `Settings.get()` return the
correct value is a lot more error-prone than modifying the read
algorithm.)

This commit also adds text to README.md about the new project-specific
settings feature.